Demo (Fine Tuning & Training)
Simulation Demo of Fine Tuning & Manual Training of the Model (local storage used for recording the successful attempts to make model learn how to get over with the traffic cars and road borders):Rate of Mutation of the Network is increased by 0.1 and Number of Genrated Cars with Randomly Created Weights & Biases are increased 10 times (to 1000) for decreasing the training time like increasing the learning rate of the model. Number of N which can be optimized by user with respect  to the hardware conditions being used at the runtime.